% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, t)
}
\arguments{
\item{N}{the number of samples}

\item{t}{the number of between-sample random numbers}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,5)
}
}
