% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtion1.R
\name{qua95}
\alias{qua95}
\title{the 0.95 quantile of a new statistic T}
\usage{
qua95(N)
}
\arguments{
\item{N}{the number of simulation}
}
\value{
the 0.95 quantile of T
}
\description{
We use the empirical distribution of a new statistic to calculate its 0.95 quantile.
}
\examples{
\dontrun{
 t <- qua95(100)
}
}
