\name{StatComp21015-package}
\alias{StatComp21015-package}
\docType{package}
\title{
  Establish my packages for 'Statistical Computing' course
}
\description{
Establish my packages for 'Statistical Computing' course, including the community Detection algorithm.
}
\details{
 Establish my packages for 'Statistical Computing' course, including the community Detection algorithm.
}
\author{
Xiyue Zhu
Maintainer: zhuxiyue@mail.ustc.edu.cn
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
