% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funtion1.R
\name{subg}
\alias{subg}
\title{perceived adjacency matrix}
\usage{
subg(A, L)
}
\arguments{
\item{A}{the observed adjacency matrix}

\item{L}{depth of knowledge}
}
\value{
the perceived adjacency matrix based on one individual
}
\description{
When we have a adjacency matrix and we are only interested in one individual, we can use establish a matrix depending on her.
}
